changed defaults and architectures
